Added loop in Factory
