+++ src/main/java/Basket/Basket.java
@@ -1,4 +1,4 @@
-package factory;
+package Basket;
 
 public class Basket {
     public void showWhatInsideBasket(){
