+++ src/main/java/Basket/RemoveProduct.java
@@ -1,4 +1,4 @@
-package factory;
+package Basket;
 
 public class RemoveProduct {
     public void removeProductFromBasket(){
