--- src/main/java/products/ProductsList.java
@@ -1,5 +1,0 @@
-package products;
-
-public interface ProductsList {
-
-}
